Move language files to separate directories
