+++ extensions/ql-vscode/src/model-editor/languages/ruby/index.ts
@@ -1,6 +1,6 @@
-import { ModelsAsDataLanguage } from "./models-as-data";
-import { sharedExtensiblePredicates, sharedKinds } from "./shared";
-import { Mode } from "../shared/mode";
+import { ModelsAsDataLanguage } from "../models-as-data";
+import { sharedExtensiblePredicates, sharedKinds } from "../shared";
+import { Mode } from "../../shared/mode";
 
 function parseRubyMethodFromPath(path: string): string {
   const match = path.match(/Method\[([^\]]+)].*/);
